# docs/swagger.yaml
basePath: /v1
definitions:
...
  dtos.AdminRegistrationRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      password_confirm:
        minLength: 6
        type: string
      secret_key:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - password_confirm
    - secret_key
    type: object
...
  dtos.CustomerRegistrationRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      password_confirm:
        minLength: 6
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - password_confirm
    type: object
  dtos.ErrorResponse:
    properties:
      error:
        example: Validation failed
        type: string
    type: object
  dtos.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword
        type: string
    required:
    - email
    - password
    type: object
  dtos.LoginSuccessResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      msg:
        example: Login successful
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/dtos.UserDetails'
    type: object
...
  dtos.RegistrationSuccessResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      msg:
        example: Account registered successfully
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/dtos.UserDetails'
    type: object
...
  dtos.UserDetails:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_login:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample e-commerce API documentation.
  title: E-Commerce API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Allows a user to login by providing email and password.
      parameters:
      - description: Login details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dtos.LoginSuccessResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: User login
      tags:
      - Auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Allows a user to register as a customer by providing necessary
        details.
      parameters:
      - description: Customer registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.CustomerRegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered customer
          schema:
            $ref: '#/definitions/dtos.RegistrationSuccessResponse'
        "400":
          description: Validation error or mismatched passwords
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Register a new customer
      tags:
      - Auth
  /register/admin:
    post:
      consumes:
      - application/json
      description: 'Allows a user to register as an admin by providing the necessary
        details, including a secret key for authentication (```secret_key: admin123```).'
      parameters:
      - description: Admin registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.AdminRegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account registered successfully
          schema:
            $ref: '#/definitions/dtos.RegistrationSuccessResponse'
        "400":
          description: Validation error or mismatched passwords
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Register a new admin
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"